@@ -42,10 +42,11 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <java.version>11</java.version>
+        <java.version>17</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
-        <testcontainers.version>1.15.1</testcontainers.version>
+        <testcontainers.version>1.19.0</testcontainers.version>
+        <hibernate.version>6.2.8.Final</hibernate.version>
     </properties>
 
     <distributionManagement>
