@@ -1,10 +1,7 @@
 package me.paulbares.domain;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import jakarta.persistence.*;
+
 
 @Entity
 @Table(name = "customer")
