@@ -8,6 +8,8 @@ import org.springframework.test.context.ContextConfigurationAttributes;
 import org.springframework.test.context.ContextCustomizer;
 import org.springframework.test.context.ContextCustomizerFactory;
 import org.springframework.test.context.MergedContextConfiguration;
+import org.testcontainers.containers.JdbcDatabaseContainer;
+import org.testcontainers.junit.jupiter.Container;
 
 import java.lang.reflect.Field;
 import java.util.List;
@@ -32,29 +34,29 @@ public class InjectableContextCustomizerFactory implements ContextCustomizerFact
     @Override
     public void customizeContext(ConfigurableApplicationContext context, MergedContextConfiguration mergedConfig) {
       List<Field> fields = ReflectionUtils.findFields(testClass,
-              f -> ReflectionUtils.isStatic(f) && f.isAnnotationPresent(AnnotationContextCustomizer.class),
+              f -> ReflectionUtils.isStatic(f) && f.isAnnotationPresent(Container.class),
               ReflectionUtils.HierarchyTraversalMode.TOP_DOWN);
-      if (fields.size() == 0) {
-        logger.debug("Cannot find a static field annotated with " + AnnotationContextCustomizer.class + " in class " + this.testClass);
+      if (fields.isEmpty()) {
+        logger.debug("Cannot find a static field annotated with " + Container.class + " in class " + this.testClass);
 
         if (this.testClass.isAnnotationPresent(SpringContainerTestDatabase.class)) {
           // Likely a config problem.
           throw new RuntimeException("It seems you are trying to use " + SpringContainerTestDatabase.class.getSimpleName()
-                  + " without adding a " + ContextCustomizer.class.getSimpleName()
+                  + " without adding a " + JdbcDatabaseContainer.class.getSimpleName()
                   + " static field annotated with "
-                  + AnnotationContextCustomizer.class.getSimpleName() + ". You should check your configuration.");
+                  + Container.class.getSimpleName() + ". You should check your configuration.");
         }
         return;
       } else if (fields.size() > 1) {
         throw new RuntimeException("Expected only 1 static field annotated with "
-                + AnnotationContextCustomizer.class + " in class " + this.testClass);
+                + Container.class + " in class " + this.testClass);
       }
 
       try {
         Field field = fields.get(0);
         field.setAccessible(true);
-        ContextCustomizer contextCustomizer = ContextCustomizer.class.cast(field.get(null));
-        contextCustomizer.customizeContext(context, mergedConfig);
+        JdbcDbContainerContextCustomizer customizer = new JdbcDbContainerContextCustomizer((JdbcDatabaseContainer) field.get(null));
+        customizer.customizeContext(context, mergedConfig);
       } catch (IllegalAccessException e) {
         throw new RuntimeException(e);
       }
